requirements.txt
customtkinter~=5.2.2
jdatetime~=5.0.0
pillow~=10.4.0